Binding bug fixed + added converters support
1) Binding now has full converters support. 7 converters are buit in.  It uses ConvertibleBinding and ConvertiblePropertyChangedHandler.
2) Bug fixed in PropertyChangedHandler.
3) DirectProperty now can be created by getter.
3) TextureLoader removed.

+++ Sgl/Base/Threading/ThreadPool.cpp
@@ -12,7 +12,7 @@ namespace Sgl
     public:
         using Task = Action<>;
 
-        ThreadPoolImpl(int maxWorkers)
+        ThreadPoolImpl(size_t maxWorkers)
         {
             _workers.reserve(maxWorkers);
